feat : Naver Login 삭제
네아로SDK 초기화 이슈로 네이버 소셜 로그인 적용시 앱 중단되는 문제 발생
카카오 로그인 하나만 써도 충분할 것 같아서 우선 네이버 소셜 로그인 로직 및 의존성 제거
