+++ app/build.gradle
@@ -72,8 +72,6 @@ dependencies {
 
     // OAuth2 Library
     implementation "com.kakao.sdk:v2-user:2.15.0" // 카카오 로그인
-    implementation 'com.google.android.gms:play-services-auth:20.6.0' // 구글 로그인
-    implementation 'com.navercorp.nid:oauth-jdk8:5.7.0' // jdk 8
 
     // Retrofit2
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
