+++ app/src/main/java/com/demo/desh/ui/theme/Type.kt
@@ -2,11 +2,9 @@ package com.demo.desh.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import com.demo.desh.R
 import customFontFamily
 
 // Set of Material typography styles to start with
@@ -19,6 +17,7 @@ val Typography = Typography(
         letterSpacing = 0.5.sp
     )
 )
+
 val Typography2 = Typography(
     bodyMedium = TextStyle(
         fontFamily = customFontFamily,
